requirements.txt
python-telegram-bot
qbittorrent-api
PyYAML
requests